+++ packages/react-components/react-carousel-preview/library/src/components/Carousel/useCarouselContextValues.ts
@@ -4,7 +4,7 @@ import type { CarouselContextValues } from '../CarouselContext.types';
 import type { CarouselState } from './Carousel.types';
 
 export function useCarouselContextValues_unstable(state: CarouselState): CarouselContextValues {
-  const { store, selectPageByDirection, selectPageByValue, circular, peeking } = state;
+  const { store, selectPageByDirection, selectPageByValue, circular } = state;
 
   const carousel = React.useMemo(
     () => ({
@@ -12,9 +12,8 @@ export function useCarouselContextValues_unstable(state: CarouselState): Carouse
       selectPageByDirection,
       selectPageByValue,
       circular,
-      peeking,
     }),
-    [store, selectPageByDirection, selectPageByValue, circular, peeking],
+    [store, selectPageByDirection, selectPageByValue, circular],
   );
 
   return { carousel };