+++ packages/react-components/react-field/library/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-field
 
-This log was last generated on Mon, 01 Jul 2024 20:30:50 GMT and should not be manually modified.
+This log was last generated on Mon, 15 Jul 2024 17:25:29 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.1.70](https://github.com/microsoft/fluentui/tree/@fluentui/react-field_v9.1.70)
+
+Mon, 15 Jul 2024 17:25:29 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-field_v9.1.69..@fluentui/react-field_v9.1.70)
+
+### Patches
+
+- fix: revert incorectly set npm versions in all packages ([PR #31937](https://github.com/microsoft/fluentui/pull/31937) by martinhochel@microsoft.com)
+- Bump @fluentui/react-context-selector to v9.1.64 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.41 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-label to v9.1.73 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-utilities to v9.18.12 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
 
 ## [9.1.69](https://github.com/microsoft/fluentui/tree/@fluentui/react-field_v9.1.69)