+++ packages/react-components/react-progress/library/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-progress
 
-This log was last generated on Mon, 01 Jul 2024 20:30:27 GMT and should not be manually modified.
+This log was last generated on Mon, 15 Jul 2024 17:25:42 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.1.81](https://github.com/microsoft/fluentui/tree/@fluentui/react-progress_v9.1.81)
+
+Mon, 15 Jul 2024 17:25:42 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-progress_v9.1.80..@fluentui/react-progress_v9.1.81)
+
+### Patches
+
+- fix: revert incorectly set npm versions in all packages ([PR #31937](https://github.com/microsoft/fluentui/pull/31937) by martinhochel@microsoft.com)
+- Bump @fluentui/react-field to v9.1.70 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.41 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.19.1 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-utilities to v9.18.12 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
 
 ## [9.1.80](https://github.com/microsoft/fluentui/tree/@fluentui/react-progress_v9.1.80)