+++ packages/react-components/react-textarea/library/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-textarea
 
-This log was last generated on Thu, 04 Jul 2024 15:14:36 GMT and should not be manually modified.
+This log was last generated on Mon, 15 Jul 2024 17:25:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.3.82](https://github.com/microsoft/fluentui/tree/@fluentui/react-textarea_v9.3.82)
+
+Mon, 15 Jul 2024 17:25:19 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-textarea_v9.3.81..@fluentui/react-textarea_v9.3.82)
+
+### Patches
+
+- fix: revert incorectly set npm versions in all packages ([PR #31937](https://github.com/microsoft/fluentui/pull/31937) by martinhochel@microsoft.com)
+- Bump @fluentui/react-field to v9.1.70 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.19.1 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-utilities to v9.18.12 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.41 ([PR #31998](https://github.com/microsoft/fluentui/pull/31998) by beachball)
 
 ## [9.3.81](https://github.com/microsoft/fluentui/tree/@fluentui/react-textarea_v9.3.81)